requirements.txt
python-socketio
aiohttp
langchain==0.0.206